% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{elasticities}
\alias{elasticities}
\alias{elasticities.quaidsce}
\title{Extract Demand Elasticities}
\usage{
elasticities(object, ...)

\method{elasticities}{quaidsce}(
  object,
  type = c("all", "income", "uncompensated", "compensated"),
  ...
)
}
\arguments{
\item{object}{A \code{quaidsce} object.}

\item{...}{Additional arguments.}

\item{type}{Character vector indicating which elasticities to return:
\code{"all"} (default), \code{"income"} (or expenditure), \code{"uncompensated"} (Marshallian),
or \code{"compensated"} (Hicksian).}
}
\value{
A list or matrix of elasticity estimates at sample means.
}
\description{
Extract Demand Elasticities
}
