% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{print.summary.quaidsce}
\alias{print.summary.quaidsce}
\title{Print Method for summary.quaidsce}
\usage{
\method{print}{summary.quaidsce}(x, digits = 4, ...)
}
\arguments{
\item{x}{A \code{summary.quaidsce} object.}

\item{digits}{Number of digits to print.}

\item{...}{Additional arguments.}
}
\description{
Print Method for summary.quaidsce
}
